requirements.txt
aniso8601
asgiref
atomicwrites
attrs
black
CacheControl
cachy
certifi
cffi
charset-normalizer
cleo
click
colorama
coreapi
coreschema
coverage
crashtest
cryptography
defusedxml
distlib
Django
django-cors-headers
django-filter
django-phonenumber-field
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
docker
drf-nested-routers
drf-spectacular
drf-yasg
dulwich
envsubst
filelock
Flask
Flask-Cors
flask-restx
Flask-SQLAlchemy
greenlet
gunicorn
html5lib
idna
importlib-metadata
inflection
iniconfig
isort
itsdangerous
itypes
jaraco.classes
Jinja2
jsonschema
jwt
keyring
lockfile
MarkupSafe
marshmallow
method-decorator
more-itertools
msgpack
mypy
mypy-extensions
oauthlib
packaging
pathspec
pexpect
phonenumbers
Pillow
pkginfo
platformdirs
pluggy
poetry
poetry-core
poetry-plugin-export
psycopg2
psycopg2-binary
ptyprocess
py
pycparser
PyJWT
pylev
pyparsing
pyrsistent
pytest
pytest-cov
pytest-env
python-dotenv
python3-openid
pytz
pywin32
pywin32-ctypes
PyYAML
requests
requests-oauthlib
requests-toolbelt
ruamel.yaml
ruamel.yaml.clib
shellingham
six
social-auth-app-django
social-auth-core
source
SQLAlchemy
sqlparse
tomli
tomlkit
types-click
types-Flask
types-Jinja2
types-MarkupSafe
types-Werkzeug
typing_extensions
tzdata
uritemplate
urllib3
virtualenv
webencodings
websocket-client
Werkzeug
xlwt
zipp
